reuse convertToScale func
