+++ src/containers/shared/utils.js
@@ -379,3 +379,17 @@ export const renderXRP = (d, language) => {
   const options = { ...CURRENCY_OPTIONS, currency: 'XRP' }
   return localizeNumber(d, language, options)
 }
+
+// Convert scaled price (assetPrice) in hex string to original price using formula:
+// originalPrice = assetPrice / 10**scale
+// More details: https://github.com/XRPLF/XRPL-Standards/tree/master/XLS-47d-PriceOracles
+export function convertScaledPrice(assetPrice, scale) {
+  const scaledPriceInBigInt = BigInt(`0x${assetPrice}`)
+  const divisor = BigInt(10 ** scale)
+  const integerPart = scaledPriceInBigInt / divisor
+  const remainder = scaledPriceInBigInt % divisor
+  const fractionalPart = (remainder * BigInt(10 ** scale)) / divisor
+  return fractionalPart > 0
+    ? `${integerPart}.${fractionalPart.toString().padStart(scale, '0')}`
+    : `${integerPart}`
+}
